Created init-scheduler script to be used to schedule initializaing an entire s3 path via a cron job
Created Docker build scripts to for the init-scheduler script

The init-scheduler depends on the init.py script so I renamed the init folder to batch-init so it now contains both the init.py Docker and the init-scheduler.py.

Started the README
