+++ docker/batch-init/.gitignore
@@ -1,0 +1,4 @@
+run.sh
+run-scheduler.sh
+build.sh
+build-scheduler.sh
