+++ docker/batch-init/build.sh.example
