+++ docker/batch-init/run.sh.example
