--- docker/init/.gitignore
@@ -1,2 +1,0 @@
-run.sh
-build.sh
